fix: making advisories packages alive
Please do not merge this commit. It works fine with an empty/small
database, but using our dump is unfeasible. I have no idea how long it
would take. I stopped execution after 5 to 8 minutes.

Just sharing that, by looking at the affected tables in
https://gist.github.com/helio-frota/45951d27442738b9cf8a43ec737fb674
it would be possible to create a bridge between the tables to
replicate the behavior of sbom_package_purl_ref, marking the advisory
packages as "alive".